Reorganised current tests to reflect new settings panel tab structure
